--- a/src/Microsoft.PowerShell.Commands.Utility/commands/utility/WebCmdlet/JsonObject.cs
+++ b/src/Microsoft.PowerShell.Commands.Utility/commands/utility/WebCmdlet/JsonObject.cs
@@ -477,12 +477,6 @@ namespace Microsoft.PowerShell.Commands
             return result;
         }
 
-        /*
-        private static object ProcessValue (object obj)
-        {
-            if (obj is )
-        }
-        */
         #endregion ConvertFromJson
 
         #region ConvertToJson