--- a/src/Microsoft.PowerShell.Commands.Utility/commands/utility/WebCmdlet/JsonObject.cs
+++ b/src/Microsoft.PowerShell.Commands.Utility/commands/utility/WebCmdlet/JsonObject.cs
@@ -263,7 +263,6 @@ namespace Microsoft.PowerShell.Commands
                 default:
                     return thisObject;
             }
-            
         }
 
         // This function is a clone of PopulateFromDictionary using JObject as an input.